+++ CHANGELOG.md
@@ -6,12 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- "CHANGELOG.md" file (issue [eclipse/keyple#6]).
+- `CHANGELOG.md` file (issue [eclipse/keyple#6]).
+- `RemotePluginServerFactoryBuilder.builder(String remotePluginName, ExecutorService executorService)` method.
 - CI: Forbid the publication of a version already released (issue [#3])
+### Upgraded
+- "Keyple Distributed Remote API" to version `2.1.0`
+- "Google Gson Library" (com.google.code.gson) to version `2.10.1`
 
 ## [2.0.0] - 2021-10-06
 This is the initial release.
-It follows the extraction of Keyple 1.0 components contained in the `eclipse/keyple-java` repository to dedicated repositories.
+It follows the extraction of Keyple 1.0 components contained in the `eclipse/keyple-java` repository to dedicated 
+repositories.
 It also brings many major API changes.
 
 [unreleased]: https://github.com/eclipse/keyple-distributed-remote-java-lib/compare/2.0.0...HEAD